[Refactor] Migrate tablet_info into StoragePrimitve
