+++ be/src/exec/pipeline/fetch_sink_operator.h
@@ -16,8 +16,8 @@
 
 #include "common/status.h"
 #include "exec/pipeline/operator_factory.h"
-#include "exec/tablet_info.h"
 #include "runtime/lookup_stream_mgr.h"
+#include "storage/primitive/tablet_info.h"
 
 namespace starrocks::pipeline {
