+++ be/src/runtime/local_tablets_channel.cpp
@@ -36,7 +36,6 @@
 #include "common/config_ingest_fwd.h"
 #include "common/statusor.h"
 #include "common/util/table_metrics.h"
-#include "exec/tablet_info.h"
 #include "gen_cpp/internal_service.pb.h"
 #include "gutil/ref_counted.h"
 #include "gutil/strings/join.h"
@@ -51,6 +50,7 @@
 #include "serde/protobuf_serde.h"
 #include "storage/delta_writer.h"
 #include "storage/memtable.h"
+#include "storage/primitive/tablet_info.h"
 #include "storage/segment_flush_executor.h"
 #include "storage/segment_replicate_executor.h"
 #include "storage/storage_engine.h"
