+++ be/src/storage/dictionary_cache_manager.cpp
@@ -18,9 +18,9 @@
 #include "column/chunk_factory.h"
 #include "column/chunk_schema_helper.h"
 #include "compute_env/dictionary_cache/chunk_util.h"
-#include "exec/tablet_info.h"
 #include "runtime/descriptors.h"
 #include "storage/chunk_helper.h"
+#include "storage/primitive/tablet_info.h"
 
 namespace starrocks {
