+++ be/src/storage/table_reader.cpp
@@ -20,13 +20,13 @@
 #include "base/brpc/ref_count_closure.h"
 #include "common/brpc/brpc_stub_cache.h"
 #include "common/brpc/internal_service_recoverable_stub.h"
-#include "exec/tablet_info.h"
 #include "platform/platform_env.h"
 #include "runtime/current_thread.h"
 #include "runtime/descriptors.h"
 #include "serde/protobuf_serde.h"
 #include "storage/chunk_helper.h"
 #include "storage/local_tablet_reader.h"
+#include "storage/primitive/tablet_info.h"
 #include "storage/storage_engine.h"
 #include "storage/tablet.h"
 #include "storage/tablet_manager.h"
