+++ be/test/exec/data_sinks/tablet_sink_test.cpp
@@ -28,12 +28,12 @@
 #include "common/config_scan_io_fwd.h"
 #include "compute_env/global_dict/fragment_dict_state.h"
 #include "compute_env/load_path/base_load_path_mgr.h"
-#include "exec/tablet_info.h"
 #include "runtime/chunk_helper.h"
 #include "runtime/descriptor_helper.h"
 #include "runtime/descriptors.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
+#include "storage/primitive/tablet_info.h"
 #include "types/decimalv2_value.h"
 
 namespace starrocks {
