+++ be/test/exec/pipeline/fetch_task_test.cpp
@@ -29,12 +29,12 @@
 #define private public
 #include "exec/pipeline/fetch_processor.h"
 #undef private
-#include "exec/tablet_info.h"
 #include "gen_cpp/Descriptors_types.h"
 #include "gtest/gtest.h"
 #include "platform/platform_env.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
+#include "storage/primitive/tablet_info.h"
 
 namespace starrocks::pipeline {
 namespace {
