+++ be/test/storage/dictionary_cache_manager_test.cpp
@@ -23,11 +23,11 @@
 #include "base/testutil/assert.h"
 #include "column/chunk_factory.h"
 #include "compute_env/dictionary_cache/chunk_util.h"
-#include "exec/tablet_info.h"
 #include "exprs/dictionary_get_expr.h"
 #include "exprs/mock_vectorized_expr.h"
 #include "runtime/descriptor_helper.h"
 #include "storage/chunk_helper.h"
+#include "storage/primitive/tablet_info.h"
 #include "storage/storage_engine.h"
 #include "storage/tablet_manager.h"
 #include "testutil/column_test_helper.h"
