+++ be/test/storage_primitive/CMakeLists.txt
@@ -32,6 +32,7 @@ set(STORAGE_PRIMITIVE_TEST_FILES
     row_source_mask_test.cpp
     storage_value_types_test.cpp
     tablet_column_test.cpp
+    tablet_info_test.cpp
     tablet_column_util_test.cpp
     tablet_column_type_utils_test.cpp
     union_iterator_test.cpp
